refactor(client): Modal 위치 변경
